requirements.txt
django==2.0
pytz==2017.2
psycopg2==2.7.7
djangorestframework==3.9.2
